segregated events (video_events.py) from gui components (video_frame) and connected it to a view_main which now creates the root Tk instance
